requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrdict==2.0.1
attrs==23.2.0
bitarray==2.9.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cytoolz==0.12.3
eth-account==0.11.0
eth-hash==0.7.0
eth-keyfile==0.8.0
eth-keys==0.5.0
eth-rlp==1.0.1
eth-typing==4.0.0
eth-utils==4.0.0
eth_abi==5.0.0
exceptiongroup==1.2.0
fastapi==0.110.0
frozenlist==1.4.1
h11==0.14.0
hexbytes==0.3.1
hiredis==2.3.2
idna==3.6
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lru-dict==1.2.0
MarkupSafe==2.1.5
multidict==6.0.5
parsimonious==0.9.0
pillow==10.2.0
protobuf==4.25.3
pycryptodome==3.20.0
pydantic==2.6.2
pydantic_core==2.16.3
pyunormalize==15.1.0
redis==5.0.2
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rlp==4.0.0
rpds-py==0.18.0
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
tabulate==0.9.0
toolz==0.12.1
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.27.1
web3==6.15.1
websockets==12.0
yarl==1.9.4
